allow py2exe find resources files plus error handling

+++ lib/epubqfix.py
@@ -23,16 +23,21 @@ from lib.htmlconstants import entities
 from lib.hyphenator import Hyphenator
 from os.path import expanduser
 
-dic_tmp_dir = tempfile.mkdtemp(suffix='', prefix='quiris-tmp-')
-dic_name = os.path.join(dic_tmp_dir, 'hyph_pl_PL.dic')
-try:
-    with open(dic_name, 'wb') as f:
-        data = get_data('lib', 'resources/dictionaries/hyph_pl_PL.dic')
-        f.write(data)
-    hyph = Hyphenator(dic_name)
-finally:
-    shutil.rmtree(dic_tmp_dir)
-
+if not hasattr(sys, 'frozen'):
+    dic_tmp_dir = tempfile.mkdtemp(suffix='', prefix='quiris-tmp-')
+    dic_name = os.path.join(dic_tmp_dir, 'hyph_pl_PL.dic')
+    try:
+        with open(dic_name, 'wb') as f:
+            data = get_data('lib', 'resources/dictionaries/hyph_pl_PL.dic')
+            f.write(data)
+        hyph = Hyphenator(dic_name)
+    finally:
+        shutil.rmtree(dic_tmp_dir)
+else:
+    hyph = Hyphenator(os.path.join(
+        os.path.dirname(sys.executable), 'resources',
+        'dictionaries', 'hyph_pl_PL.dic'
+    ))
 MY_LANGUAGE = 'pl'
 HYPHEN_MARK = u'\u00AD'
 
@@ -338,8 +343,14 @@ def fix_html_toc(soup, tempdir, xhtml_files, xhtml_file_paths):
             )
         else:
             parser = etree.XMLParser(remove_blank_text=True)
-            transform = etree.XSLT(etree.fromstring(get_data('lib',
-                                   'resources/ncx2end-0.2.xsl')))
+            if not hasattr(sys, 'frozen'):
+                transform = etree.XSLT(etree.fromstring(get_data('lib',
+                                       'resources/ncx2end-0.2.xsl')))
+            else:
+                transform = etree.XSLT(etree.parse(os.path.join(
+                    os.path.dirname(sys.executable), 'resources',
+                    'ncx2end-0.2.xsl'
+                )))
             toc_ncx_file = etree.XPath(
                 '//opf:item[@media-type="application/x-dtbncx+xml"]',
                 namespaces=OPFNS
@@ -774,8 +785,13 @@ def replace_svg_html_cover(opftree, rootepubdir):
                                   parser=etree.XMLParser(recover=True))
     svg_imgs = html_cover_tree.xpath('//svg:image', namespaces=SVGNS)
     if len(svg_imgs) == 1:
-        new_cover_tree = etree.fromstring(get_data('lib',
-                                          'resources/cover.xhtml'))
+        if not hasattr(sys, 'frozen'):
+            new_cover_tree = etree.fromstring(get_data('lib',
+                                              'resources/cover.xhtml'))
+        else:
+            new_cover_tree = etree.parse(os.path.join(
+                os.path.dirname(sys.executable), 'resources', 'cover.xhtml'
+            ))
         new_cover_tree.xpath(
             '//xhtml:img',
             namespaces=XHTMLNS
